requirements.txt
streamlit==1.25.0
openai==0.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
langchain-openai==0.0.5
langchain-community==0.0.5
